Adding some package level logging and some whitespace fixup.
