+++ pypd/models/vendor.py
@@ -1,3 +1,4 @@
+
 # Copyright (c) PagerDuty.
 # See LICENSE for details.
 from .entity import Entity
