+++ pypd/version.py
@@ -1,4 +1,6 @@
+
 # Copyright (c) PagerDuty.
 # See LICENSE for details.
-version = (0, 0, 1)
+
+version = (0, 1, 0)
 __version__ = '.'.join(map(str, version))
